Start to fix Attractors example (#408)
* explicitly require array to have float dtype

* same change for lorenz and rossler

+++ envisage/examples/demo/plugins/tasks/attractors/model/rossler.py
@@ -1,4 +1,5 @@
 # System library imports.
+import numpy as np
 from scipy import arange, array
 from scipy.integrate import odeint
 
@@ -40,7 +41,7 @@ class Rossler(HasTraits):
     c = Float(5.7, auto_set=False, enter_set=True)
 
     # Integration parameters.
-    initial_point = Array(value=[0.0, 1.0, 0.0])
+    initial_point = Array(dtype=np.float64, value=[0.0, 1.0, 0.0])
     time_start = Float(0.0)
     time_stop = Float(100.0)
     time_step = Float(0.01)
